first commit, clientmax body fix, potential time out done

+++ src/WebServ/WebServerProg.hpp
@@ -30,7 +30,7 @@ struct Server
 	std::string serverName;
     std::map<int, std::string> errorPages;
     std::vector<Location> locations; 
-    int clientMaxBodySize;
+    double clientMaxBodySize;
 	std::string uploadDirectory;
 };
